style: clear the 141 non-doc clippy warnings from toolchain drift
Triage first. NONE of the three suspect classes turned out to be a real
defect; each was checked, not assumed:

* `precedence` x6 (range_coder.rs x4, pd0.rs, qm.rs) -- every one is
  `a * b >> c`. Rust binds `*` tighter than `>>`, so the expression
  ALREADY evaluates as clippy's suggested `(a * b) >> c`, and C binds
  them the same way, so the ported form was faithful. Parenthesized for
  clarity; zero behaviour change (byteid gate confirms).
* `identity_op` x3 -- all in test code. intrabc_hash.rs x2 open-coded
  `add_value` for block_size 4 as a baked `+ (0 << 16)` / `+ 0`; both now
  derive it as `hash_block_size_to_index(4).unwrap() << CRC_BITS`, the
  way production and the third test site already do, so the tests track
  the real key composition. quant.rs's `>> 0` was documenting
  `>> log_scale` at log_scale 0 (c_tx_size 2); now a named binding.
* `manual_div_ceil` x32 -- `(x + 1) / 2` on `usize` frame dimensions at
  every site (checked each: pipeline true/upscaled/4x4 dims, and test /
  example chroma dims). `x + 1` can only overflow at `usize::MAX`, which
  no frame dimension reaches, so `div_ceil(2)` is equal on the whole
  reachable domain -- and strictly better outside it. pipeline.rs:4010
  already used `div_ceil(2)` two lines below one of them.

Also checked and NOT a bug: `if_same_then_else` in
`intra_edge_filter_strength` -- C intra_prediction.c:190-197 really does
have two identical `d >= 40` arms (blk_wh <= 12 and <= 16). Kept
shape-faithful behind a targeted allow citing the C line.

Fixed properly (no suppression): div_ceil x32, unnecessary_cast x14
(all genuinely same-type; var_boost's was load-bearing for INFERENCE, so
the literals gained `f64` suffixes instead), int_plus_one x4,
manual_range_contains x4, manual_range_patterns x4, needless_borrow x3,
duplicated_attributes x3, question_mark, derivable_impls,
field_reassign_with_default x2, unnecessary_unwrap-adjacent cleanups,
useless_vec, useless_conversion, cloned_ref_to_slice_refs,
print_with_newline, 2 unused imports, and items_after_test_module x2
(the `mod tests` blocks in dsp/cdef.rs and encoder/var_boost.rs moved to
the end of their files, so the production code after them is visible).

`leaf_funnel.rs:5676` `lane.len() <= cap - 1` -> `< cap` is the one
conversion with an overflow question: `cap` is `... + 1` at both call
sites so `cap >= 1` always and the two forms agree.

Targeted `#[allow]` with a reason at each site, no crate-level blanket:
* type_complexity x9 -- ported C signatures / inline tuple tables where
  a `type` alias hides the shape.
* manual_checked_ops x6 -- the `> 0` guard scopes a whole block, not one
  division; `checked_div` cannot express it without restructuring hot RD
  control flow.
* unnecessary_unwrap x3 -- the `if let` rewrite needs a let-chain
  (Rust 1.88) and this workspace declares `rust-version = "1.85"`.
* dead_code x2 -- `build_refined_scan` is a faithful port of C
  `perform_pred_depth_refinement`; rust/CLAUDE.md forbids deleting
  dead-looking C. `y_recon`/`uv_recon` round out a used accessor family.
* large_enum_variant, single_element_loop x2, if_same_then_else x1.

Gates: byteid_fingerprint cells=168, port==C on 168, 0 rows moved vs
benchmarks/byteid_fingerprint_2026-08-11.tsv. nextest 1029/1029 release
AND 1029/1029 debug. `cargo +stable fmt --all --check` clean.
`#![forbid(unsafe_code)]` proved live: a planted
`unsafe { transmute::<u32, f32> }` in svtav1-encoder fails the build
("usage of an `unsafe` block"), removed after.

+++ rust/crates/svtav1-dsp/tests/quantize_neon_parity.rs
@@ -77,7 +77,7 @@ fn quantize_matches_scalar_across_shapes_and_divisors() {
 
 #[test]
 fn quantize_handles_zero_dequant() {
-    let coeffs: Vec<i32> = (0..64).map(|i| (i as i32) * 37 - 900).collect();
+    let coeffs: Vec<i32> = (0..64).map(|i| i * 37 - 900).collect();
     for &(dc, ac) in &[(0i32, 24i32), (20, 0), (0, 0)] {
         let qp = QuantParam {
             dequant: [dc, ac],
@@ -98,10 +98,10 @@ fn quantize_falls_back_past_the_f32_exact_bound() {
     // anything bigger tests nothing but the test's own arithmetic. With
     // shift=4 and dequant=24: coeff 1<<22 gives shifted 1<<26 (past the bound,
     // so the fallback runs) and q*dequant = 67,108,848, well inside i32.
-    let under: Vec<i32> = (0..64).map(|i| 1 << 15 | (i as i32)).collect();
-    let over: Vec<i32> = (0..64).map(|i| (1 << 22) + (i as i32) * 1013).collect();
+    let under: Vec<i32> = (0..64).map(|i| 1 << 15 | i).collect();
+    let over: Vec<i32> = (0..64).map(|i| (1 << 22) + i * 1013).collect();
     let mixed: Vec<i32> = (0..64)
-        .map(|i| if i == 31 { 1 << 22 } else { i as i32 * 11 })
+        .map(|i| if i == 31 { 1 << 22 } else { i * 11 })
         .collect();
     for &shift in &[0i32, 2, 4] {
         let qp = QuantParam {
